refactor: 디자인 시스템 컬러 패키지 변경
